Crate axlibc

ArceOS user program library for C apps.

Cargo Features

fd: Enable file descriptor table.

pipe: Enable pipe support.

select: Enable synchronous I/O multiplexing (select) support.

epoll: Enable event polling (epoll) support.

Functions

ax_accept⚠net

Accept for connections on a socket

ax_bind⚠net

Bind a address to a socket.

ax_clock_gettime⚠

Get clock time since booting

ax_close⚠fd

Close a file by fd.

ax_connect⚠net

Connects the socket to the address specified.

ax_dup⚠fd

Duplicate a file descriptor

ax_dup3⚠fd

dup3() is the same as dup2(), except that:

ax_epoll_create⚠epoll and (select or epoll)

ax_epoll_create() creates a new epoll instance.

ax_epoll_ctl⚠epoll and (select or epoll)

Control interface for an epoll file descriptor

ax_epoll_wait⚠epoll and (select or epoll)

ax_epoll_wait() waits for events on the epoll instance referred to by the file descriptor epfd.

ax_errno_string⚠

Returns a pointer to the string representation of the given error code.

ax_exit⚠

Exits the current thread.

ax_fcntl⚠fd

Fcntl implementation

ax_free⚠alloc

Deallocate memory.

ax_fstat⚠fd

Get file metadata by fd and write into buf.

ax_getaddrinfo⚠net

Query addresses for a domain name.

ax_getcwd⚠fs

Get the path of the current directory.

ax_getpeername⚠net

Get peer address to which the socket sockfd is connected.

ax_getpid⚠multitask

Get current thread ID.

ax_getsockname⚠net

Get current address to which the socket sockfd is bound.

ax_listen⚠net

Listen for connections on a socket

ax_lseek⚠fs

Set the position of the file indicated by fd.

ax_lstat⚠fs

Get the metadata of the symbolic link and write into buf.

ax_malloc⚠alloc

Allocate memory and return the memory address.

ax_mktime⚠

mktime implementation

ax_nanosleep⚠

Sleep some nanoseconds

ax_open⚠fs

Open a file by filename and insert it into the file descriptor table.

ax_panic⚠

Abort the current process.

ax_pipe⚠pipe

Create a pipe

ax_print_str⚠

Print a string to the global standard output stream.

ax_println_str⚠

Print a string to the global standard output stream. Add a line break.

ax_pthread_create⚠multitask

Create a new thread with the given entry point and argument.

ax_pthread_exit⚠multitask

Exits the current thread. The value retval will be returned to the joiner.

ax_pthread_join⚠multitask

Waits for the given thread to exit, and stores the return value in retval.

ax_pthread_mutex_init⚠multitask

Initialize a mutex.

ax_pthread_mutex_lock⚠multitask

Lock the given mutex.

ax_pthread_mutex_unlock⚠multitask

Unlock the given mutex.

ax_rand_u32⚠

Returns a 32-bit unsigned pseudo random interger.

ax_read⚠fd

Read data from the file indicated by fd.

ax_recv⚠net

Receive a message on a socket.

ax_recvfrom⚠net

Receive a message on a socket and get its source address.

ax_select⚠select and (select or epoll)

Monitor multiple file descriptors, waiting until one or more of the file descriptors become “ready” for some class of I/O operation

ax_send⚠net

Send a message on a socket to the address connected.

ax_sendto⚠net

Send a message on a socket to the address specified.

ax_shutdown⚠net

Shut down a full-duplex connection.

ax_socket⚠net

Create an socket for communication.

ax_srand⚠

Sets the seed for the random number generator.

ax_stat⚠fs

Get the file metadata by path and write into buf.

ax_strftime⚠alloc

strftime implementation

ax_strtod⚠fp_simd

strtod implementation

ax_strtof⚠fp_simd

strtofimplementation

ax_sysconf⚠

Return system configuration infomation

ax_write⚠fd

Write data to the file indicated by fd.

ax_writev⚠fd

writev implementation